feat: 주문서 작성 -> 주문 dto 작성 중

+++ src/main/java/com/nhnacademy/order_payments/dto/cart/BookApiResponse.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.order_payments.dto.cart;
 
 public record BookApiResponse (
-        long bookId,
+        Long bookId,
         String title,
-        int price
+        Long price
 ) {}
