+++ src/main/java/com/nhnacademy/order_payments/dto/response/order/OrderDetailResponse.java
@@ -6,7 +6,7 @@ public record OrderDetailResponse(
         long orderDetailId,
         String bookTitle,
         String bookCoverImage,
-        int price, //구매당시 금액
+        Long price, //구매당시 금액
         int quantity,
         String packagingName, // 포장을 안했으면 null, 했으면 포장한 포장지 이름
         OrderDetailStatus orderDetailStatus,
